+++ buildSrc/src/main/groovy/org/apache/beam/gradle/BeamModulePlugin.groovy
@@ -1532,33 +1532,25 @@ class BeamModulePlugin implements Plugin<Project> {
           def disabledChecks = [
             // TODO(https://github.com/apache/beam/issues/20955): Enable errorprone checks
             "AutoValueImmutableFields",
-            "AutoValueImmutableFields",
             "AutoValueSubclassLeaked",
             "BigDecimalEquals",
             "ComparableType",
             "DoNotMockAutoValue",
             "EmptyBlockTag",
-            "EqualsGetClass",
             "EqualsUnsafeCast",
             "EscapedEntity",
             "ExtendsAutoValue",
             "InlineMeSuggester",
             "InvalidBlockTag",
-            "InvalidThrows",
             "JavaTimeDefaultTimeZone",
-            "JavaUtilDate",
             "JodaConstructors",
             "MixedMutabilityReturnType",
             "PreferJavaTimeOverload",
             "NonCanonicalType",
-            "Slf4jFormatShouldBeConst",
             "Slf4jSignOnlyFormat",
             "ThreadPriorityCheck",
-            "TimeUnitConversionChecker",
             "UndefinedEquals",
             "UnescapedEntity",
-            "UnnecessaryMethodReference",
-            "UnnecessaryParentheses",
             "UnrecognisedJavadocTag",
             // errorprone 3.2.0+ checks
             "DirectInvocationOnMock",
@@ -1569,7 +1561,6 @@ class BeamModulePlugin implements Plugin<Project> {
             "NullableTypeParameter",
             "NullableWildcard",
             "SuperCallToObjectMethod",
-            "UnusedVariable",
             // intended suppressions emerged in newer protobuf versions
             // For backward compatibility. Public method checked in before this check impl
             // Possible use in interface subclasses
@@ -2322,7 +2313,7 @@ class BeamModulePlugin implements Plugin<Project> {
 
       // This sets the whole project Go version.
       // The latest stable Go version can be checked at https://go.dev/dl/
-      project.ext.goVersion = "go1.26.0"
+      project.ext.goVersion = "go1.26.1"
 
       // Minor TODO: Figure out if we can pull out the GOCMD env variable after goPrepare script
       // completion, and avoid this GOBIN substitution.