+++ examples/java/src/main/java/org/apache/beam/examples/RegexExample.java
@@ -76,7 +76,7 @@ public class RegexExample {
 
     @ProcessElement
     public void processElement(ProcessContext c) throws Exception {
-      LOG.info(prefix + c.element());
+      LOG.info("{}{}", prefix, c.element());
       c.output(c.element());
     }
   }