+++ sdks/java/io/amazon-web-services2/src/main/java/org/apache/beam/sdk/io/aws2/schemas/AwsSchemaProvider.java
@@ -134,7 +134,7 @@ public class AwsSchemaProvider extends GetterBasedSchemaProviderV2 {
       if (this == o) {
         return true;
       }
-      if (o == null || getClass() != o.getClass()) {
+      if (!(o instanceof FromRowWithBuilder)) {
         return false;
       }
       FromRowWithBuilder<?> that = (FromRowWithBuilder<?>) o;