+++ sdks/java/io/google-cloud-platform/src/main/java/org/apache/beam/sdk/io/gcp/bigtable/BigtableIO.java
@@ -1545,7 +1545,7 @@ public class BigtableIO {
           String.format(
               "At least %d errors occurred writing to Bigtable. First %d errors: %s",
               i + failures.size(), i, logEntry.toString());
-      LOG.error(message);
+      LOG.error("{}", message);
       IOException exception = new IOException(message);
       for (BigtableWriteException e : suppressed) {
         exception.addSuppressed(e);
@@ -2462,7 +2462,7 @@ public class BigtableIO {
       // Only try to create or update metadata table if option is set to true. Otherwise, just
       // check if the table exists.
       if (shouldCreateOrUpdateMetadataTable && metadataTableAdminDao.createMetadataTable()) {
-        LOG.info("Created metadata table: " + metadataTableId);
+        LOG.info("Created metadata table: {}", metadataTableId);
       }
     }
 
@@ -2645,7 +2645,7 @@ public class BigtableIO {
       // Only try to create or update metadata table if option is set to true. Otherwise, just
       // check if the table exists.
       if (metadataTableAdminDao.createMetadataTable()) {
-        LOG.info("Created metadata table: " + metadataTableAdminDao.getTableId());
+        LOG.info("Created metadata table: {}", metadataTableAdminDao.getTableId());
       }
       return metadataTableAdminDao.doesMetadataTableExist();
     } finally {