+++ sdks/java/io/google-cloud-platform/src/main/java/org/apache/beam/sdk/io/gcp/testing/FakeDatasetService.java
@@ -19,8 +19,8 @@ package org.apache.beam.sdk.io.gcp.testing;
 
 import static org.junit.Assert.assertEquals;
 
-import com.google.api.client.googleapis.json.GoogleJsonResponseException;
 import com.google.api.client.http.HttpHeaders;
+import com.google.api.client.http.HttpResponseException;
 import com.google.api.core.ApiFuture;
 import com.google.api.core.ApiFutures;
 import com.google.api.gax.grpc.GrpcStatusCode;
@@ -703,7 +703,7 @@ public class FakeDatasetService implements DatasetService, WriteStreamService, S
             if (this.updatedSchema == null) {
               this.updatedSchema = newSchema;
               this.schemaInformation =
-                  TableRowToStorageApiProto.SchemaInformation.fromTableSchema((this.updatedSchema));
+                  TableRowToStorageApiProto.SchemaInformation.fromTableSchema(this.updatedSchema);
             }
           }
         }
@@ -811,7 +811,7 @@ public class FakeDatasetService implements DatasetService, WriteStreamService, S
   void throwNotFound(@FormatString String format, Object... args) throws IOException {
     throw new IOException(
         String.format(format, args),
-        new GoogleJsonResponseException.Builder(404, String.format(format, args), new HttpHeaders())
+        new HttpResponseException.Builder(404, String.format(format, args), new HttpHeaders())
             .build());
   }
 }